requirements.txt
asgiref==3.8.1
chardet==5.2.0
cssselect2==0.7.0
Django==4.2.16
django-admin-sortable2==2.2.3
django-filer==3.2.3
django-polymorphic==3.1.0
easy-thumbnails==2.10
lxml==5.3.0
pillow==11.0.0
python-dotenv==1.0.1
reportlab==4.2.5
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
webencodings==0.5.1
